+++ otri/downloader/yahoo_downloader.py
@@ -21,18 +21,13 @@ from . import (ATOMS_KEY, META_KEY_DOWNLOAD_DT, META_KEY_EXPIRATION,
                META_TS_VALUE_TYPE, METADATA_KEY, OptionsDownloader,
                TimeseriesDownloader)
 
-META_VALUE_PROVIDER = "yahoo finance"
-
 
 class YahooTimeseries(TimeseriesDownloader):
     '''
     Used to download historical time series data from YahooFinance.
     '''
 
-    def __init__(self):
-        # Import meta provider value to have it externally available
-        global META_VALUE_PROVIDER
-        YahooTimeseries.META_VALUE_PROVIDER = META_VALUE_PROVIDER
+    META_VALUE_PROVIDER = "yahoo finance"
 
     def history(self, ticker: str, start: date, end: date, interval: str = "1m", max_attempts: int = 5) -> Union[dict, bool]:
         '''
@@ -116,7 +111,7 @@ class YahooTimeseries(TimeseriesDownloader):
         data[METADATA_KEY] = {
             META_KEY_TICKER: ticker,
             META_KEY_INTERVAL: interval,
-            META_KEY_PROVIDER: META_VALUE_PROVIDER,
+            META_KEY_PROVIDER: YahooTimeseries.META_VALUE_PROVIDER,
             META_KEY_TYPE: META_TS_VALUE_TYPE,
             META_KEY_DOWNLOAD_DT: th.now()
         }
@@ -210,7 +205,7 @@ class YahooOptions(OptionsDownloader):
         # Append medatada values
         chain[METADATA_KEY] = {
             META_KEY_TICKER: ticker,
-            META_KEY_PROVIDER: META_VALUE_PROVIDER,
+            META_KEY_PROVIDER: YahooTimeseries.META_VALUE_PROVIDER,
             META_KEY_OPTION_TYPE: kind,
             META_KEY_DOWNLOAD_DT: th.now(),
             META_KEY_EXPIRATION: expiration,
@@ -373,6 +368,6 @@ class YahooMetadata:
             except Exception as e:
                 log.e("there has been an exception when retrieving ticker {} ISIN: {}".format(ticker, e))
             # Add provider
-            info['provider'] = [META_VALUE_PROVIDER]
+            info['provider'] = [YahooTimeseries.META_VALUE_PROVIDER]
             data.append(info)
         return data
